requirements.txt
streamlit
numpy
matplotlib
Pillow
yaml
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
torchvision==0.3.0